Integração

+++ weagle-app/app/src/main/java/com/fiapos/weagle/features/ideas/presentation/edit/EditIdeaViewModel.kt
@@ -31,9 +31,7 @@ class EditIdeaViewModel(
 
         viewModelScope.launch {
 
-            idea = repository.getIdeaById(
-                ideaId.toInt()
-            )
+            idea = repository.getIdeaById(ideaId)
         }
     }
