+++ weagle-app/app/src/main/java/com/fiapos/weagle/features/so/presentation/edit/EditStrategicOrientationViewModel.kt
@@ -32,9 +32,7 @@ class EditStrategicOrientationViewModel(
 
         viewModelScope.launch {
 
-            orientation = repository.getOrientationById(
-                orientationId.toInt()
-            )
+            orientation = repository.getOrientationById(orientationId)
         }
     }
